feat: 메인화면 수정 & 기능 구현

+++ src/main/java/com/trendhub/trendhub/domain/coordi/repository/CoordiRepository.java
@@ -3,5 +3,5 @@ package com.trendhub.trendhub.domain.coordi.repository;
 import com.trendhub.trendhub.domain.coordi.entity.Coordi;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CoordiRepository extends JpaRepository<Coordi, Integer>, CoordiRepositoryCustom {
+public interface CoordiRepository extends JpaRepository<Coordi, Long>, CoordiRepositoryCustom {
 }
